use setup to test waiting for players
